+++ src/frontend/apps/standalone_site/src/features/Profile/components/InfoSettings.tsx
@@ -1,0 +1,5 @@
+import { Box } from 'grommet';
+
+export const InfoSettings = () => {
+  return <Box></Box>;
+};